fix data class
